Refactor init() and show welcome message last in main() startup to make start cleaner
